Add X86 CPU name and feature detection
Also implement the API needed by other part of the runtime,
especially LLVM JIT, cloning pass and sysimg initialization.

Most of the detection (both name and features) code is copied from LLVM.
However, it's very hard for us to use LLVM because there are informations we need that
is not provided by LLVM (some of which is non-trivial to expose in a target independent way),
including,

1. Feature dependencies
2. Features related to a CPU name
3. Feature name validation (needed when we expose this to julia code)
4. Effect on ABI (vector register size)
5. Serialization format
6. Cloning heuristic

Additionally, the detection code itself is only a small part of the code for each arch
and we need to support multiple LLVM versions so copying the LLVM code shouldn't cause too
much problem by itself.